pyproject.toml
[project]
name = "interview-mocker-backend"
version = "1.0.0"
description = "AI模拟面试与能力提升平台 - 后端服务"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    # FastAPI Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # Database
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2",
    # Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    # Vector Database (for RAG)
    "chromadb>=0.4.22",
    # LangChain (for RAG)
    "langchain>=0.1.4",
    "langchain-community>=0.0.16",
    "sentence-transformers>=2.3.1",
    # Document Processing
    "pypdf>=4.0.1",
    "python-docx>=1.1.0",
    "markdown>=3.5.2",
    # Redis (optional for production)
    "redis>=5.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["main_service", "interview_avatar_service", "shared"]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["main_service", "interview_avatar_service", "shared"]
skip = [".git", ".venv", "venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
